Refactoring of the language component #70
