+++ src/components/pages/UserMasterPasswordValidation/index.js
@@ -41,6 +41,7 @@ import NotLogged from '../../templates/NotLogged/index.js';
 import withUser from '../../../providers/WithUser.js';
 // Hooks
 import useDialogConfirmation from '../../../hooks/useDialogConfirmation/index.js';
+import AtomLanguageSelector from '../../atoms/LanguageSelector/index.js';
 
 const PageUserMasterPasswordValidation = ({user}) => {
     const { t } = useTranslation();
@@ -112,6 +113,7 @@ const PageUserMasterPasswordValidation = ({user}) => {
             </ButtonWrapper>
             
             <AtomButtonLink onClick={logout}>{t('common.Logout')}</AtomButtonLink>
+            <AtomLanguageSelector />
         </NotLogged>
     </>
